WEB-itinvent/backend/requirements.txt
fastapi
uvicorn[standard]
pyodbc
python-jose
pydantic[email]
python-dotenv
python-docx
docx2pdf
openpyxl
openai
pypdf
reportlab
PyMuPDF
pysnmp
cryptography
exchangelib
sqlalchemy
alembic
psycopg[binary]
pywebpush
pyotp
qrcode
redis
slowapi
webauthn
